refactor(mcp-server): split canvas.ts route file into cohesive sub-routers
The 853-line canvas.ts mixed workspace/canvas CRUD, name/pin metadata,
live-doc snapshot+update, version history, thumbnails, restore, and
maintenance (compact/prune/optimize) in one file. Split each concern
into its own module under routes/canvas/, composed by routes/canvas.ts,
so each is independently readable and testable. Behavior-preserving:
same route paths, status codes, and response bodies; existing
canvas.test.ts and route-coverage.test.ts pass unchanged.
